# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BBS论坛用户管理API
  description: |
    BBS论坛系统的用户认证与管理接口文档
    
    ## 功能概述
    本API提供用户注册、登录、个人信息管理等核心功能，支持多种认证方式。
    
    ## 认证机制
    - JWT Token认证：登录成功后返回JWT Token，需在请求头中携带 `Authorization: Bearer <token>`
    - 三种登录方式：
      1. 用户名密码登录：传统的用户名+密码认证
      2. 邮箱验证码登录：通过邮箱接收验证码进行认证
      3. 微信小程序登录：使用微信小程序授权码登录
    - 用户注册支持：邮箱验证码注册、微信小程序注册
    
    ## 错误处理
    所有接口均返回统一格式的响应，包含状态码、消息和数据。错误时会返回对应的错误码和错误信息。
  version: 0.1.0
  contact:
    name: BBS开发团队-后端组
    email: leotanstudio@gmail.com
servers:
  - url: http://bbs.leocoding.onliine
    description: 测试环境
paths:
...
  /user/{username}:
    get:
      tags:
        - 用户管理
      summary: 根据用户名获取用户信息
      description: |
        获取指定用户名的用户公开信息。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            example: "johndoe"
          description: 用户名
      responses:
        '200':
          description: 获取用户信息成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posts/draft:
    get:
      tags:
        - 帖子管理
      summary: 获取当前用户帖子草稿
      description: |
        获取当前登录用户最新保存的帖子草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 获取草稿成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/draft/{draftId}:
    put:
      tags:
        - 帖子管理
      summary: 编辑帖子草稿
      description: |
        更新指定草稿的标题、正文或分类信息。
        
        **权限要求：** 需要登录且必须是草稿作者
      security:
        - bearerAuth: [ ]
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: integer
          description: 草稿ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftContentEditorRequest'
      responses:
        '200':
          description: 草稿更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限编辑该草稿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/draft/{draftId}/publish:
    post:
      tags:
        - 帖子管理
      summary: 发布帖子草稿
      description: |
        将指定草稿发布为公开帖子。
        
        **权限要求：** 需要登录且必须是草稿作者
      security:
        - bearerAuth: [ ]
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: integer
          description: 草稿ID
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: 草稿状态不允许发布或内容存在敏感风险
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限发布该草稿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts:
    get:
      tags:
        - 帖子管理
      summary: 分页查询公开帖子
      description: |
        获取已发布帖子列表，支持按分类筛选与分页。
        
        **默认排序：** 热度优先
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            example: "学习心得"
          description: 按分类过滤帖子
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPageResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - 帖子管理
      summary: 直接发布帖子
      description: |
        直接发布一个仅包含文字内容的帖子，跳过草稿箱流程。

        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: 请求参数错误或内容存在敏感风险
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/hot:
    get:
      tags:
        - 帖子管理
      summary: 获取热门帖子
      description: |
        获取系统推荐的热门帖子列表，支持分页。
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPageResponse'
  /posts/search:
    get:
      tags:
        - 帖子管理
      summary: 搜索帖子
      description: |
        根据关键词搜索已发布帖子，支持模糊匹配。
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
            example: "计算机网络"
          description: 搜索关键词
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPageResponse'
        '400':
          description: 搜索条件无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
        description: 帖子ID
    get:
      tags:
        - 帖子管理
      summary: 获取帖子详情
      description: |
        根据帖子ID查看已发布帖子的完整信息。
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: 帖子不存在或未发布
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 帖子管理
      summary: 删除帖子
      description: |
        删除指定帖子。
        
        **权限要求：** 需要登录且必须是帖子作者或具有管理权限
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限删除该帖子
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}/toggle-status:
    patch:
      tags:
        - 帖子管理
      summary: 切换帖子上下线状态
      description: |
        在上线与下线状态间切换帖子，常用于隐藏或恢复帖子。
        
        **权限要求：** 需要登录且必须是帖子作者
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 状态切换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限执行该操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}/to-draft:
    patch:
      tags:
        - 帖子管理
      summary: 将帖子转为草稿
      description: |
        将已发布的帖子重新保存为草稿状态，用户可继续编辑。
        
        **权限要求：** 需要登录且必须是帖子作者
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 转换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限执行该操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}/publish:
    patch:
      tags:
        - 帖子管理
      summary: 发布已存在的帖子
      description: |
        将草稿或下线的帖子重新发布。
        
        **权限要求：** 需要登录且必须是帖子作者
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: 帖子状态不允许发布或内容存在敏感风险
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限执行该操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/user/{userId}:
    get:
      tags:
        - 帖子管理
      summary: 查看用户发布的帖子
      description: |
        根据用户ID分页查询该用户发布的公开帖子。
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPageResponse'
...
  /comments/post/{postId}:
    get:
      tags:
        - 评论管理
      summary: 分页获取帖子评论
      description: |
        获取帖子下的评论列表（包含互动信息），支持分页。
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,asc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPageResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Token认证，格式为：Bearer <token>
        
        获取方式：
        1. 用户名密码登录：POST /auth/login
        2. 邮箱验证码登录：POST /auth/login-with-code
        3. 微信小程序登录：POST /auth/login-with-wechat
        
        Token会在登录成功后通过响应头Authorization和响应体data.token同时返回。
  schemas:
...
    UserDetails:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          type: object
          properties:
            id:
              type: integer
              description: 用户ID
              example: 123
            username:
              type: string
              description: 用户名
              example: "johndoe"
            nickname:
              type: string
              description: 昵称
              example: "John Doe"
            bio:
              type: string
              description: 个人简介
              example: "Hello, I'm John!"
            profileSlug:
              type: string
              description: 个人页面地址
              example: "my-profile"
            avatarUrl:
              type: string
              description: 头像URL
              example: "https://storage.example.com/avatar/123.jpg"
            role:
              type: string
              enum:
                - USER
                - ADMIN
                - MODERATOR
              description: 用户角色
              example: "USER"
            status:
              type: string
              enum:
                - Active
                - Inactive
                - Banned
              description: 用户状态
              example: "Active"
            verifiedId:
              type: integer
              description: 认证信息ID
              example: 456
            createdAt:
              type: string
              format: date-time
              description: 创建时间
              example: "2024-01-01T00:00:00Z"
            updatedAt:
              type: string
              format: date-time
              description: 更新时间
              example: "2024-01-01T12:00:00Z"
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误代码
          example: 10001
        message:
          type: string
          description: 错误消息
          example: "Invalid parameters"
        data:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties:
                type: string
          description: 错误详细信息，验证失败时包含字段级错误
          example:
            username: "用户名长度必须在2-50个字符之间"
            password: "密码长度必须在8-20个字符之间"
    DraftContentEditorRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: 帖子标题
          example: "如何准备计算机网络考试"
        content:
          type: string
          description: 帖子正文内容（可为空表示继续保留原内容）
          example: "分享一些复习计算机网络的资料和经验..."
        category:
          type: string
          maxLength: 100
          description: 帖子分类
          example: "学习心得"
    PostAuthor:
      type: object
      description: 帖子作者信息
      properties:
        id:
          type: integer
          example: 1001
        username:
          type: string
          example: "johndoe"
        nickname:
          type: string
          example: "约翰"
        bio:
          type: string
          example: "喜欢分享编程经验"
        profileSlug:
          type: string
          example: "john-profile"
        avatarUrl:
          type: string
          format: uri
          example: "https://cdn.example.com/avatar/john.jpg"
        role:
          type: string
          enum:
            - USER
            - ADMIN
            - MODERATOR
          example: "USER"
        status:
          type: string
          enum:
            - Active
            - Inactive
            - Banned
          example: "Active"
        verifiedId:
          type: integer
          nullable: true
          example: 10
        createdAt:
          type: string
          format: date-time
          example: "2024-02-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-05-01T12:30:00Z"
    Medium:
      type: object
      description: 媒体资源信息
      properties:
        id:
          type: string
          example: "media_abc123"
        displayUrl:
          type: string
          format: uri
          description: 用于前端展示的访问地址
          example: "https://cdn.example.com/post/media_abc123.jpg"
        resourceUrl:
          type: string
          format: uri
          description: 资源真实访问地址
          example: "https://storage.example.com/media/media_abc123"
        type:
          type: string
          description: 媒体类型，如image/video/cover等
          example: "image"
    PostDetail:
      type: object
      description: 帖子详细信息
      properties:
        id:
          type: integer
          example: 2001
        author:
          $ref: '#/components/schemas/PostAuthor'
        title:
          type: string
          example: "计算机网络复习笔记"
        content:
          type: string
          example: "这是一个关于计算机网络的详细复习资料..."
        media:
          type: array
          items:
            $ref: '#/components/schemas/Medium'
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - SUSPENDED
            - UNDER_REVIEW
          example: "PUBLISHED"
        possiblySensitive:
          type: boolean
          example: false
        category:
          type: string
          example: "学习心得"
        createdAt:
          type: string
          format: date-time
          example: "2024-05-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-05-02T08:30:00Z"
    SortInfo:
      type: object
      description: 排序信息
      properties:
        sorted:
          type: boolean
          example: false
        unsorted:
          type: boolean
          example: true
        empty:
          type: boolean
          example: true
    PageMetadata:
      type: object
      description: 分页元数据
      properties:
        sort:
          $ref: '#/components/schemas/SortInfo'
        pageNumber:
          type: integer
          description: 当前页码（从0开始）
          example: 0
        pageSize:
          type: integer
          description: 每页数量
          example: 10
        offset:
          type: integer
          description: 偏移量
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false
    PostPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostDetail'
        pageable:
          $ref: '#/components/schemas/PageMetadata'
        totalElements:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 5
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        sort:
          $ref: '#/components/schemas/SortInfo'
        numberOfElements:
          type: integer
          example: 10
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
    PostResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          $ref: '#/components/schemas/PostDetail'
    PostPageResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          $ref: '#/components/schemas/PostPage'
    BooleanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          type: boolean
          example: true
...
    CreatePostRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 255
          description: 帖子标题
          example: "校园学习经验分享"
        content:
          type: string
          description: 帖子正文内容
          example: "今天整理了几条备考建议，大家可以交流一下。"
        category:
          type: string
          maxLength: 100
          nullable: true
          description: 帖子分类，可选
          example: "学习心得"
...
    CommentDetail:
      type: object
      description: 评论详情（包含互动信息）
      properties:
        id:
          type: integer
          example: 3001
        postId:
          type: integer
          example: 2001
        userId:
          type: integer
          example: 1001
        username:
          type: string
          example: "johndoe"
        userAvatar:
          type: string
          format: uri
          example: "https://cdn.example.com/avatar/john.jpg"
        parentCommentId:
          type: integer
          nullable: true
          example: null
        content:
          type: string
          example: "非常感谢分享！"
        media:
          type: array
          items:
            $ref: '#/components/schemas/Medium'
        comments:
          type: integer
          description: 回复数量
          example: 2
        likes:
          type: integer
          description: 点赞数量
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2024-05-01T12:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-05-02T09:10:00Z"
        isLiked:
          type: boolean
          description: 当前用户是否已点赞
          example: true
...
    CommentPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CommentDetail'
        pageable:
          $ref: '#/components/schemas/PageMetadata'
        totalElements:
          type: integer
          example: 24
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        sort:
          $ref: '#/components/schemas/SortInfo'
        numberOfElements:
          type: integer
          example: 8
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
    CommentPageResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          $ref: '#/components/schemas/CommentPage'
...